feat: add site configuration and type definitions for site components
